Downloads/scitechjournals/requirements.txt
asgiref
Django==5.0.1
psycopg2-binary
requests
openpyxl
python-dotenv
requests
beautifulsoup4
cloudscraper
selenium
pandas
psycopg2
webdriver-manager
regex
panda
django-cors-headers
django-environ
djangorestframework
celery
redis
django-celery-results
undetected-chromedriver
pycountry
pyinstaller
tkinterweb
pystray
pillow
python-dotenv
lxml
rapidfuzz
pydrive
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib